Fix follower re-start bug which caused only last state machine action to be sent
Caused by shadowing the data pointer in the leader, rather than copying it in.
(This commit introduces a hideous memory leak -- not too much of a bother for the
purposes this code is used for, but disastrous for production code.  It is unclear
at present how this should be resolved within the constraints of the engine.)